Fixed bug involving distances below offset
Previously, code may panic if there are pair
distances below the lower limit of the lowest bin,
due to the way the binary search algorithm was written.
Rewrote to guarantee that r passed to the binary search
algorithm is in some bin, which greatly simplifies the
logic and makes us more confident that it is correct. I believe
that r above the upper limit was not a problem because used
floor in selecting ties for next checked bin.